Port to 64bit & CBash v0.7.1
Squashed version of a bunch of commits to keep them from breaking dev
and to minimize the number of commits containing changes to CBash.dll
(since it's binary).

CBash v0.7.1

This is the 64bit version, built using zlib v1.2.11 and boost 1.71.0.

Default Oblivion Patch now runs on 64bit

Tested with Oblivion and its DLCs and on the default settings for the
bashed patch. Further work is obviously needed but this serves as a good
staging point to understand what needs to change.

Added static typing to ctypes - performance will probably take a hit,
but did not profile either version. This lets us spot errors much easier
though, which is thoroughly worth it for this stage of the process.

First - there are tons of type errors in the ffi with CBash. Unsigned
cast as signed, mixing None and 0, etc. It seems(?) to run fine on 32bit
though, so not sure if it should be fixed.

Second - the problem lies squarely with those 'c_ulong' that Infernio
identified before. Some are in the functions he spotted, but there are
others spread about (in this run, the glaring ones are under those
'_ParentID' methods).

Re: c_void_p - it almost certainly works since we don't really care
about the contents themselves, but using those opaque structs like in
this commit should be prefered - even if we remove static typing on
arguments, using them for return types gives us a bit more safety, more
readability and shouldn't impact performance.

'All' Oblivion patchers now run on 64bit

Tested with Oblivion and its DLCs with 'Select All' patchers.

On the python only a few mistyped arguments needed to change while on
the C++ side a union was being initialized with the expectation that
int's size was the same as a pointer's. These changes are in the
'ganda-wip' branch in CBash and the compiled Debug dll is included in
this commit.

All Oblivion patchers now run on 64bit

Tested with the following load order:
00  Oblivion.esm
01  All Natural Base.esm  [Version 1.38]
02  Cobl Main.esm  [Version 1.72]
03  Cobl Glue.esp  [Version 1.72]
++  Cobl Tweaks.esp  [Version 1.44]
04  Unofficial Oblivion Patch.esp  [Version 3.5.6]
05  UOP Vampire Aging & Face Fix.esp  [Version 1.0.0]
++  Oblivion Citadel Door Fix.esp
06  DLCShiveringIsles.esp
07  Cobl Si.esp  [Version 1.63]
08  DLCHorseArmor.esp
09  DLCOrrery.esp
0A  DLCVileLair.esp
0B  DLCMehrunesRazor.esp
0C  DLCSpellTomes.esp
0D  DLCThievesDen.esp
0E  DLCBattlehornCastle.esp
0F  DLCFrostcrag.esp
10  Knights.esp
11  All Natural - Real Lights.esp  [Version 1.38]
12  TIE.esp  [Version 1.49]
13  All Natural - SI.esp  [Version 1.38]
14  All Natural.esp  [Version 1.38]
15  Roads of Cyrodiil.esp  [Version 1.0]
++  RealWalk.esp
16  Salmo the Baker, Cobl.esp  [Version 3.08]
17  Cobl Races.esp  [Version 1.52]
++  Oblivion DLC Delayers.esp  [Version 1.0.1]
++  ROC Standard Road Record.esp
++  ROC - OCReborn Road Record.esp
++  Cobl Races - Balanced.esp  [Version 1.52]
++  Cobl Filter Late MERGE ONLY.esp  [Version 1.53]
18  DragonInvasion.esp
19  Bashed Patch, 0.esp

Don't think I forgot anything and neither should any patcher be silently
skipped.

Fix TaskDialogIndirect calls failing on 64bit

It was a packing issue because Windows is stupid.

Fix building for 64bit

Moved all packages to 64bit (only the wheels and loot needed to be
changed).

Changed the manifest to look for the 64bit dlls.

Removed the two 64bit dll exclusions from setup.py

Exclude the two new DLLs that now get included

Not sure if this is needed, but those DLLs now show up in the built EXE,
and one of them seems to be MSVC redist, so...

Fix 'Scan for Dirty Edits' erroring

I think this illustrates why we should keep the static typing, made
identifying the problem pretty simple :)
That said, my C knowledge is terrible, so this might actually be wrong.

Also dropped unused windll import.

Various readme changes, drop Gohlke links

Those are temporary, so we have to host mirrors ourselves. Obviously
temporary, which is why I kept them on a branch that we can later drop.
Edited the readmes to drop the '32bit is needed' notes and outdated
things (e.g. if you get a MemoryError now, something is *severely*
wrong with your system :P).

Co-authored-by: Infernio <infernio@icloud.com>