Refactor FactionRelations to use _AParser
Added some functionality to become Skyrim-compatible in the next commit,
that's why the added line count is unusually high.

See how similar _write_record is to the one from ActorFactions? Could
probably create an abstract class for this, but we'll worry about that
later - this refactoring will likely expose many such opportunities for
follow-up refactoring.