WIP use precompiled struct unpackers:
This is not so much about performance as for centralizing the handling
of packers and unpackers so we can further extend/error check etc
Note we should really only override loadData in brec - note similarity
in some of the overrides. We should also use EAFP for errors and avoid
if checks - note also using dictionaries to the same end.
Re: optimization, note that apart from using precompiled Struct, we
also bring stuff to local scope. This is done with the dunder prefixed
function parameters - they are not meant to be used by clients.
Unfortunately this introduces more `.unpack` occurrences - fortunately
you can search for the calls with (?<!\))\.unpack\b