Fix CBash crashing while trying to print errors EEE
For some reason, this method wanted to print the address of a collection
pointer. I don't see how that would *ever* be useful information, plus
it's broken, so I dropped it. Also dropped a useless encode() call at
the end of that method.

Might want to merge this with the 64bit branch, they might be related.

EEE we should really take this opportunity to rework how the BP handles
and displays errors

Closes # 486 <--- RRR