Drop removed tags stored in pickle files
If a tag gets removed but it was manually added to a mod, it will still
be there, with no way to remove it except to check 'Automatic' and redo
all manual tagging. There was actually already a constant for 'old tags'
in WB, so I simply expanded that to take all tags we've removed so far
into account (at least, these are the ones I know have been removed -
there may be more).

Note that I dropped the automatic description editing that the old code
did. There is no notification to users when it happens (and adding one
would mean more coupling between bosh and balt...), so you just suddenly
end up with BAIN-mismatched files. This is something mod authors
should fix, not WB.