SSE: Make a copy of plugins.txt for the CK
The CK looks for plugins.txt in the game folder, not AppData\Local,
which is where it *should* be looking. So make a copy of the file
whenever we write it out.

Note that this is fixed in FO4, and seems to crash Skyrim LE's CK when a
plugins.txt file is present, so we only do it for SSE.